Geo-parse text documents java file
The attached tool takes three arguments: an input file which lists all
documents to be processed, the bucket to which the output file should be
written, and the name of the resulting output file.

The output file has multiple columns and is TAB-DELIMITED.  These
columns are Document, Location, Latitude, Longitude, and Context.
Document is the source-document of the change, Location is the
Gazeteer's rendering of the location (a cleaned name), Latitutde is the
location's latitude, longitude is the location's longitude, and context
is the neighboring text of the location, both before and afterwards, a
maximum of 150 characters.