Updated CalculateDistance Jar with point reconciliation
Points that are really close together are now treated as the same point,
this is a variable threshold based on the given threshold for edge
connectedness.  Currently, it is 1/10 of the threshold for edge
connectedness.

This is primarily to address the issue where many many points are right
on top of each other due to Gazetteer lookup, etc .