Updated location reconciliation to handle tags intelligently
Tags from discarded locations (to form the merged location) are compared
and added intelligently to the new location.

Every tag is separated by white-space, each tag is checked for whether
the current location already has the tag, and if it does, to add it in.
If not, the tag is thrown away.