Updated Java code to give clear signal of AWS config fail
Surrounded the check of buckets for output for all three AWS utilities
with a specific try/catch.  If an error is thrown here, the culprit is
the AWS configuration (probably the pass/key phrase, or maybe the bucket
ID).

Previously, the code will simply fail out with a -1.  The code now exits
immediately with a -2 after reporting those two lines to System.out